RegionCFGTree: remove costly method orderNodes
This method was used to sort vector of nodes either in post-order or in
reverse-post-order.

It had 2 problems:
- the boolean argument used to select between POT and RPOT was no longer
  used, since all the remaining call sites passed false.
- it was expensive to call, because it took a vector, it created a set
  from it, and performed a reverse-post-order-traversal of the whole
  graph while pushing the nodes to sort back into a newly allocated
  vector. This was particularly problematic because in one of the call
  sites a reverse-post-order-traversal was already available, but the
  method did calculate a new one nonetheless.

The method is removed in this commit.
All the call sites have been modified to use the
reverse-post-order-traversal object if already available, and they
avoid allocating a new set for the computation and a new vector for the
results. The initialization is now performed directly into a set, and
the sorted results are pushed into a vector. This effectively saves the
allocation of a vector in each call site.