Hack for function isolation work with symbols
Add some hacks in order to avoid errors when performing function
isolation after having elected CFPEs only with debug symbols.

This is necessary since, when using symbols, calls to external functions
that pass through the `.plt`, do not have a target (the `.plt` stub is
not elected as a function obviously).