DLACreateIntraProcedualTypes: create less pointers
Before this commit, the creation of layouts in the DLA was very
aggressive, treating almost every instruction as if it could be an
address (hence creating a layout for it).

This commit adds assertions and narrows down the number of cases where
an instruction is actually considered an address.
Specifically, the following kinds of instructions are not considered
addresses anymore:
  - Mul
  - SDiv
  - UDiv
  - SRem
  - URem
  - AShr
  - LShr
  - Shl
  - And
  - Xor
  - Or