CMakeLists.txt: reduce the number of tests
Before this commit, the decompilation tests were executed on all the
artifacts coming from revng runtime tests.
This commit restricts this, running it only on the artifacts for the
'printf' runtime test.

We consider this restriction not to be harmful to tests, because the
'printf' test already contains all the libc, which is actually what
really stress-tests the decompiler.
Running only on 'printf' should be enough to trigger bugs on rare code
paths, as showed in the past, while considerably reducing the runtime of
the tests.