Tests: workaround a bug of revng arg --prefix
The --prefix option of revng has a bug that causes it to skip over the
whole command line, ignoring all the other arguments, and always
printing the help returning with success.
This means that we cannot properly detect failing tests.

Luckyly, if we use the syntax `--prefix=actualprefix` instead of
`--prefix actualprefix` the bug does not show up and we can properly run
our tests.