BinaryFile.h: detect GNU_IFUNC as Unknown symbols
It is quite common for new users to compile programs with their system
compilers and feeding them to revng.
Most system compilers link C programs against glibc, which emits GNU
indirect function symbols even when compiled with -static.
Such symbols are not supported by revng yet, but until now the default
behaviour was to crash with the unhelpful message: "Unexpected symbol
type".

This commits changes the behavior for GNU_IFUNC symbols, detecting them
as Unknown symbols and reporting a warning message whenever such a
symbol is detected in the lifted binary.

This should allow the lifting process to succeed with soft errors in a
larger set of cases.