Proper Segment mapping + entrypoints optionality
We can consider an epoch to be unique. So we can search for the relati-
ve segment by passing only the epoch to the method. Right now, we will
fine the unique corrisponding segment, but in a possible future with
many epochs within a single segment, this will also works.